[DEV-9186a] Update Legend Shapes
